Move serialization primitives to System.Runtime
We need BinaryFormatter to be able to serialize types from coreclr that use the various serialization types (e.g. ISerializable), which means those types must have the same identity as the ones BinaryFormatter recognizes, and we need a bunch of types all across corefx to be able to use these same types to implement serialization.  As such, the types are moving down to System.Runtime, and on coreclr, are type-forwarding to the implementations there.

There are a few outstanding issues:
- SR is not currently available in System.Runtime.  The few uses of SR in the versions of the code used for corert have been replaced by hardcoded strings and TODOs.
- Debug.Assert is not currently available in System.Runtime.  Asserts have been commented out with TODOs.
- To get things to build as types are moving around, P2P references are used.
