+++ src/System.Runtime.Serialization.Formatters/src/System/Runtime/Serialization/ObjectManager.cs
@@ -1624,4 +1624,17 @@ namespace System.Runtime.Serialization
 
         internal string TypeName { get; }
     }
+
+    // TODO: Temporary workaround.  Remove this once SerializationInfo.UpdateValue is exposed
+    // from coreclr for use by ObjectManager.
+    internal static class SerializationInfoExtensions
+    {
+        private static readonly Action<SerializationInfo, string, object, Type> s_updateValue =
+            (Action<SerializationInfo, string, object, Type>)typeof(SerializationInfo)
+            .GetMethod("UpdateValue", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance)
+            .CreateDelegate(typeof(Action<SerializationInfo, string, object, Type>));
+
+        public static void UpdateValue(this SerializationInfo si, string name, object value, Type type) =>
+            s_updateValue(si, name, value, type);
+    }
 }
