+++ src/System.Runtime.Serialization.Formatters/tests/project.json
@@ -21,12 +21,10 @@
     "Microsoft.DotNet.xunit.performance": "1.0.0-alpha-build0037"
   },
   "frameworks": {
-    "netstandard1.3": {}
+    "netstandard1.5": {}
   },
   "supports": {
     "coreFx.Test.netcoreapp1.0": {},
-    "coreFx.Test.net46": {},
-    "coreFx.Test.net461": {},
     "coreFx.Test.net462": {},
     "coreFx.Test.net463": {}
   }
