+++ src/System.Runtime/ref/System.Runtime.cs
@@ -1377,6 +1377,13 @@ namespace System
         public static bool operator ==(System.MulticastDelegate d1, System.MulticastDelegate d2) { return default(bool); }
         public static bool operator !=(System.MulticastDelegate d1, System.MulticastDelegate d2) { return default(bool); }
     }
+    [AttributeUsage(AttributeTargets.Field, Inherited = false)]
+    public sealed class NonSerializedAttribute : Attribute
+    {
+        public NonSerializedAttribute()
+        {
+        }
+    }
     public partial class NotImplementedException : System.Exception
     {
         public NotImplementedException() { }
@@ -1546,6 +1553,13 @@ namespace System
         public static bool TryParse(string s, System.Globalization.NumberStyles style, System.IFormatProvider provider, out sbyte result) { result = default(sbyte); return default(bool); }
         [System.CLSCompliantAttribute(false)]
         public static bool TryParse(string s, out sbyte result) { result = default(sbyte); return default(bool); }
+    }
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Enum | AttributeTargets.Delegate, Inherited = false)]
+    public sealed class SerializableAttribute : Attribute
+    {
+        public SerializableAttribute()
+        {
+        }
     }
     [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
     public partial struct Single : System.IComparable, System.IComparable<float>, System.IConvertible, System.IEquatable<float>, System.IFormattable
@@ -3488,6 +3502,172 @@ namespace System.Runtime.InteropServices
         public int Size;
         public StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind layoutKind) { }
         public System.Runtime.InteropServices.LayoutKind Value { get { return default(System.Runtime.InteropServices.LayoutKind); } }
+    }
+}
+namespace System.Runtime.Serialization
+{
+    public interface IDeserializationCallback
+    {
+        void OnDeserialization(object sender);
+    }
+    [CLSCompliant(false)]
+    public interface IFormatterConverter
+    {
+        object Convert(object value, Type type);
+        object Convert(object value, TypeCode typeCode);
+        bool ToBoolean(object value);
+        char ToChar(object value);
+        [CLSCompliant(false)]
+        sbyte ToSByte(object value);
+        byte ToByte(object value);
+        short ToInt16(object value);
+        [CLSCompliant(false)]
+        ushort ToUInt16(object value);
+        int ToInt32(object value);
+        [CLSCompliant(false)]
+        uint ToUInt32(object value);
+        long ToInt64(object value);
+        [CLSCompliant(false)]
+        ulong ToUInt64(object value);
+        float ToSingle(object value);
+        double ToDouble(object value);
+        Decimal ToDecimal(object value);
+        DateTime ToDateTime(object value);
+        String ToString(object value);
+    }
+    public interface IObjectReference
+    {
+        object GetRealObject(StreamingContext context);
+    }
+    public interface ISerializable
+    {
+        void GetObjectData(SerializationInfo info, StreamingContext context);
+    }
+    [System.AttributeUsageAttribute((System.AttributeTargets)(64), Inherited = false)]
+    public sealed partial class OnDeserializedAttribute : System.Attribute
+    {
+        public OnDeserializedAttribute() { }
+    }
+    [System.AttributeUsageAttribute((System.AttributeTargets)(64), Inherited = false)]
+    public sealed partial class OnDeserializingAttribute : System.Attribute
+    {
+        public OnDeserializingAttribute() { }
+    }
+    [System.AttributeUsageAttribute((System.AttributeTargets)(64), Inherited = false)]
+    public sealed partial class OnSerializedAttribute : System.Attribute
+    {
+        public OnSerializedAttribute() { }
+    }
+    [System.AttributeUsageAttribute((System.AttributeTargets)(64), Inherited = false)]
+    public sealed partial class OnSerializingAttribute : System.Attribute
+    {
+        public OnSerializingAttribute() { }
+    }
+    [System.AttributeUsageAttribute(System.AttributeTargets.Field, Inherited = false)]
+    public sealed partial class OptionalFieldAttribute : System.Attribute
+    {
+        public OptionalFieldAttribute() { }
+        public int VersionAdded { get { return default(int); } set { } }
+    }
+    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
+    public struct SerializationEntry
+    {
+        public string Name { get { throw null; } }
+        public Type ObjectType { get { throw null; } }
+        public object Value { get { throw null; } }
+    }
+    public partial class SerializationException : System.Exception
+    {
+        public SerializationException() { }
+        public SerializationException(string message) { }
+        public SerializationException(string message, System.Exception innerException) { }
+    }
+    public sealed class SerializationInfo
+    {
+        [CLSCompliant(false)]
+        public SerializationInfo(Type type, IFormatterConverter converter) { }
+        public string AssemblyName { get { throw null; } set { } }
+        public string FullTypeName { get { throw null; } set { } }
+        public int MemberCount { get { throw null; } }
+        public Type ObjectType { get { throw null; } }
+        public bool IsFullTypeNameSetExplicit { get { throw null; } }
+        public bool IsAssemblyNameSetExplicit { get { throw null; } }
+        public void AddValue(string name, bool value) { }
+        public void AddValue(string name, byte value) { }
+        public void AddValue(string name, char value) { }
+        public void AddValue(string name, DateTime value) { }
+        public void AddValue(string name, decimal value) { }
+        public void AddValue(string name, double value) { }
+        public void AddValue(string name, short value) { }
+        public void AddValue(string name, int value) { }
+        public void AddValue(string name, long value) { }
+        public void AddValue(string name, object value) { }
+        public void AddValue(string name, object value, Type type) { }
+        [CLSCompliant(false)]
+        public void AddValue(string name, sbyte value) { }
+        public void AddValue(string name, float value) { }
+        [CLSCompliant(false)]
+        public void AddValue(string name, ushort value) { }
+        [CLSCompliant(false)]
+        public void AddValue(string name, uint value) { }
+        [CLSCompliant(false)]
+        public void AddValue(string name, ulong value) { }
+        public bool GetBoolean(string name) { throw null; }
+        public byte GetByte(string name) { throw null; }
+        public char GetChar(string name) { throw null; }
+        public DateTime GetDateTime(string name) { throw null; }
+        public decimal GetDecimal(string name) { throw null; }
+        public double GetDouble(string name) { throw null; }
+        public SerializationInfoEnumerator GetEnumerator() { throw null; }
+        public short GetInt16(string name) { throw null; }
+        public int GetInt32(string name) { throw null; }
+        public long GetInt64(string name) { throw null; }
+        [CLSCompliant(false)]
+        public sbyte GetSByte(string name) { throw null; }
+        public float GetSingle(string name) { throw null; }
+        public string GetString(string name) { throw null; }
+        [CLSCompliant(false)]
+        public ushort GetUInt16(string name) { throw null; }
+        [CLSCompliant(false)]
+        public uint GetUInt32(string name) { throw null; }
+        [CLSCompliant(false)]
+        public ulong GetUInt64(string name) { throw null; }
+        public object GetValue(string name, Type type) { throw null; }
+        public void SetType(Type type) { }
+    }
+    public sealed class SerializationInfoEnumerator : System.Collections.IEnumerator
+    {
+        private SerializationInfoEnumerator() { }
+        public SerializationEntry Current { get { throw null; } }
+        public string Name { get { throw null; } }
+        public Type ObjectType { get { throw null; } }
+        object System.Collections.IEnumerator.Current { get { throw null; } }
+        public object Value { get { throw null; } }
+        public bool MoveNext() { throw null; }
+        public void Reset() { throw null; }
+    }
+    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
+    public partial struct StreamingContext
+    {
+        public StreamingContext(System.Runtime.Serialization.StreamingContextStates state) { }
+        public StreamingContext(System.Runtime.Serialization.StreamingContextStates state, object additional) { }
+        public override bool Equals(object obj) { return default(bool); }
+        public override int GetHashCode() { return default(int); }
+        public System.Runtime.Serialization.StreamingContextStates State { get { return default(System.Runtime.Serialization.StreamingContextStates); } }
+        public object Context { get { return default(object); } }
+    }
+    [Flags]
+    public enum StreamingContextStates
+    {
+        CrossProcess = 0x01,
+        CrossMachine = 0x02,
+        File = 0x04,
+        Persistence = 0x08,
+        Remoting = 0x10,
+        Other = 0x20,
+        Clone = 0x40,
+        CrossAppDomain = 0x80,
+        All = 0xFF,
     }
 }
 namespace System.Runtime.Versioning
