+++ src/System.Runtime/src/System/NonSerializedAttribute.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
 namespace System
 {
     [AttributeUsage(AttributeTargets.Field, Inherited = false)]
