+++ src/System.Runtime/src/System/Runtime/Serialization/IObjectReference.cs
