+++ src/System.Runtime/src/System/Runtime/Serialization/ISerializable.cs
