+++ src/System.Runtime/src/System/Runtime/Serialization/OnDeserializedAttribute.cs
@@ -2,14 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-//------------------------------------------------------------
-
-using System;
-
 namespace System.Runtime.Serialization
 {
     [AttributeUsage(AttributeTargets.Method, Inherited = false)]
-    public sealed class OnSerializingAttribute : Attribute
+    public sealed class OnDeserializedAttribute : Attribute
     {
     }
 }
