+++ src/System.Runtime/src/System/SerializableAttribute.cs
